moving funcs from initMap to mapCtrler
